lr reset
